Nested serializers are read-only now
Nested serializers are complicated to work with if they are not
read-only. As this is not needed for us in this project to modify
nested serializers, they were changed to read-only.

Also user pk in the training model is now changed to the username, as
it was easier to work with API this way.